trunk, cisstDataGenerator: use size_t consistantly for line numbers
git-svn-id: https://svn.lcsr.jhu.edu/cisst/trunk@3924 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ cisst/cisstCommon/applications/cisstDataGenerator/cdgBaseClass.cpp
@@ -25,7 +25,7 @@ http://www.cisst.org/cisst/license.txt.
 CMN_IMPLEMENT_SERVICES(cdgBaseClass);
 
 
-cdgBaseClass::cdgBaseClass(unsigned int lineNumber):
+cdgBaseClass::cdgBaseClass(size_t lineNumber):
     cdgScope(lineNumber)
 {
     cdgField * field;
@@ -53,7 +53,7 @@ cdgScope::Type cdgBaseClass::GetScope(void) const
 
 bool cdgBaseClass::HasScope(const std::string & CMN_UNUSED(keyword),
                             cdgScope::Stack & CMN_UNUSED(scopes),
-                            unsigned int CMN_UNUSED(lineNumber))
+                            size_t CMN_UNUSED(lineNumber))
 {
     return false;
 }
