+++ cisst/cisstCommon/applications/cisstDataGenerator/cdgClass.h
@@ -64,12 +64,12 @@ class cdgClass: public cdgScope
     BaseClassesType BaseClasses;
 
 public:
-    cdgClass(unsigned int lineNumber);
+    cdgClass(size_t lineNumber);
 
     cdgScope::Type GetScope(void) const;
     bool HasScope(const std::string & keyword,
                   cdgScope::Stack & scopes,
-                  unsigned int lineNumber);
+                  size_t lineNumber);
     void GenerateHeader(std::ostream & outputStream) const;
     void GenerateCode(std::ostream & outputStream) const;
